Changed BibClassify ontology management.
Removed the ontology files from Invenio. The ontology file have now to be
hosted in a different place and BibClassify will download them whenever
they are updated.

This commit also fixes a bug in the regular expressions generated from the
ontology. Concepts of one character will be sensible to case.